Prevent some errors due to strings instead of arrays
selected_topic_list was being handled sometimes as a string and sometimes as an array, and sometimes not checked for string.
Changed to make it an array if it wasn't and then treat it as an array.

Removed a setting of user prefs current_problem to null that happens after session timeout which was causing a 0 prob_id in the
response table.