Some changes based on code review - split updating topic from problem update, utility to create select options from array
Update problem was updating problem and topic - split this and call two updates rather than doing it all in the problem update

Changed the utility to accept a key,value array and current value to create select options rather than doing it mixed in with
view creation in the view. So the label,values are created separately and then given to the utility to create the layout.