PR-67 show message when all problems have been completed correctly rather than skip btn
When all problems in a topic have been completed correctly, the user should get a message
saying so. Instead they were seeing a disabled submit button and an enabled skip button ,and
a blank page below.

Added a check for null to a local variable which is being used now to hold problem_id
rather than the session variable that used to be used.