PR-72 confusion about the X next to Add btns
Made the x appear only when the form for the section appears,
and added a cancel button tot he form section as well. Changed the css a
bit to make the 3 sections more distinguishable.

Made the x in the button more accessible by hiding it to screen readers
and adding for screen readers some descriptive text as to what the
button does.

Also a tiny change to improve the contrast of the footer because it
wasn't readable.