PR-98 submit rating with next button instead of submit rating button
Moved the rating div/inputs to be inside the next form instead of having their own form, and
include the ratings.php in the next post handling.

Also fixed a transpose of ave and count in the clarity rating display when editing a problem, and there
was a typo in using v_rating rather then v_ratings to cycle through ratings, so an error was shown on the
edit problem page and the rating wasn't being shown.

Added a missing check for number of tries - the rating scale should be shown if user tries >= number of
delay_solution.