PR-111 - Add votes to the REST API
There is a new migration to add the table, but votes are not yet set up
as a first class model. The only interface to them is the REST API,
which allows only a list of votes for the current user in a given
course, and creating or updating a vote for a problem as the current
user.