PR-112 - Add PRApi script and ES6 Promise polyfill
The PRApi script gives a simple JavaScript object to use to interact
with the REST API. It contains a dummy implementation so it can be used
without a live API endpoint and then configured to use AJAX without
changing the client code. All methods are exposed under the Promises/A+
API, allowing the success/failure handling to be more elegant and
chainable.