PR-122 PR-123 make problems the landing page, fix topic id for problem retry in response
Change the landing page from selections to problems for an initial workaround fix to the response
time problem when first coming into PR. Also fix the topic id logged with a response which was
incorrect when a problem was retried. Was using a get current topic which didn't really. Needed to
pass the topic in the post when a problem is retried.