PR-129 - Improve topic selection performance further
This removes the basically unused MCTSelect class, eliminating the
extraneous queries, and simplifies the VTopic_Selections view.

There is also some query profiling included, which can be turned on by
setting $GLOBALS['DEBUG'] = true in paths.inc.php. It is simple,
counting fetch_assoc invocations and each of its immediate callers.
The result is echoed in an HTML comment at the bottom of the body.