PR-135 - Fix up syntax for PHP 5.2
There was one spot in setup.php where the Caliper config was using
syntax not compatible with PHP 5.2. This is the inline
constructor/object pattern. `(new SomeClass)->method()` does not work in
PHP 5.2. The rest of the Caliper code is protected behind the enabled
flag, which will imply a compatible runtime, but this was in the code
that always gets parsed.

This commit also has some syntax fixes in lib/views.php that were
obscuring the issue. That is that some of the control structures did not
have trailing semicolons, and there were a couple of short tags used not
for echoing.