PR-137 - Expire timeouts after an hour of inactivity
This change has two components:

  1. The server keeps track of the latest activity in the session and
     redirects if a request comes in after more than an hour
  2. The browser sets a timer to redirect to the timeout page at an
     hour.

If a request is made directly to timeout.php, the current session is
terminated. The tabs are presented, so it is easy to go to any other
page in the app and activate a new session. The sessions are still not
recorded in the database.