PR-153 - Refactor MProblem constructor to not load
The constructor was always loading from the database, which made it
difficult to do efficient batch loading. This change refactors the
constructor to have no arguments and provides find/fill methods.

It also avoids loading the "OK to display solution" flag on load, since
it is only used on one page, one problem at a time. The same applies for
the topic names; they are put behind an accessor that will load and
cache the names for a given problem when needed (only on single problem
pages at present).

This is the first piece of batching the problem loading for My Summary.