PR-153 - Improve loading problems within a topic
There are two improvements here. The first is that, when returning just
IDs, the rest of the columns are no longer selected. The second, and
larger one, is that there is now only one query run, rather than one for
the IDs and then a loop to load the problems.